% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_track.r
\name{add_timestamp_to_tracks}
\alias{add_timestamp_to_tracks}
\title{add_timestamp_to_tracks
Adds date/time information to the output of read_track()
Adds columns "day", "time", and (optionally) "date" (if start_date supplied)}
\usage{
add_timestamp_to_tracks(track_df, start_date)
}
\arguments{
\item{track_df}{data.frame, output of read_track()}

\item{start_date}{(optional) a Date object indicating the start date of the track file}
}
\value{
data.frame containing all supplied columns plus day, time, date (if start_date supplied)
}
\description{
add_timestamp_to_tracks
Adds date/time information to the output of read_track()
Adds columns "day", "time", and (optionally) "date" (if start_date supplied)
}
