% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate_storm.R
\name{associate_storm}
\alias{associate_storm}
\title{associate_storm}
\usage{
associate_storm(storms, precip, interest, storm_radius, precip_threshold)
}
\arguments{
\item{storms}{data.frame, output of \code{\link[=read_track]{read_track()}}}

\item{precip}{Raster or spatRaster map of precipitation}

\item{interest}{sf Polygons object detailing the region of interest}

\item{storm_radius}{numeric, a threshold identifying how near to a precip band
a track centre must be to be considered to be associated with the band.}

\item{precip_threshold}{numeric, a threshold value identifying how intense
precipitation must be to be considered to be a "band"}
}
\value{
list containing Precip, PrecipRegions, MatchedStorm, MatchedPath,
Interest, AllStorms, StormBuffers, Plot
}
\description{
Identifies the most likely storm track point to be associated
with a weather event in a defined interest region, based on the ability of the
algorithm to draw a path through a continuous region of precipitation. The
function first identifies regions of precipitation exceeding a given
threshold. It then attempts to draw lines through these regions of
precipitation to connect the region of interest with storm track centres. If
more than one track is found to be linked to the interest region via
precipitation bands, the track joined to the interest region via the shortest
path through precipitation bands is selected.
}
\examples{
example_precip <- terra::rast(system.file("extdata", "EXAMPLE_era5_precip.tif", package = "FrontTracking"))

results <- associate_storm(
 storms = example_storms,
 precip = example_precip,
 interest = england_wales_box,
 storm_radius = 250,
 precip_threshold = 1.5
)
}
\seealso{
\code{\link[=read_track]{read_track()}} (required for loading Hodges' TRACK data)

\code{\link[terra:rast]{terra::rast()}} (required for loading NCDF precipitation data)
}
