% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_track.r
\name{convert_track_to_negative_longitude}
\alias{convert_track_to_negative_longitude}
\title{convert_track_to_negative_longitude
Converts a track's longitude from a 0:360 degree schema to -180:180 degree}
\usage{
convert_track_to_negative_longitude(track_df)
}
\arguments{
\item{track_df}{data.frame: export from read_track()}
}
\value{
data.frame: same format as original with modified longitude values
}
\description{
convert_track_to_negative_longitude
Converts a track's longitude from a 0:360 degree schema to -180:180 degree
}
