% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_track.r
\name{crop_tracks_to_extent}
\alias{crop_tracks_to_extent}
\title{crop_tracks_to_extent
Crop the output of read_track to a given spatial extent}
\usage{
crop_tracks_to_extent(
  track_df,
  extent = c(xmin = -70, xmax = 25, ymin = 20, ymax = 75)
)
}
\arguments{
\item{track_df}{data.frame, output of read_track()}

\item{extent}{numeric vector containing entries xmin, xmax, ymin, ymax}
}
\value{
data.frame
}
\description{
crop_tracks_to_extent
Crop the output of read_track to a given spatial extent
}
