% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rsmc_besttrack.R
\name{read_rsmc_besttrack}
\alias{read_rsmc_besttrack}
\title{read_rsmc_besttrack}
\usage{
read_rsmc_besttrack(file)
}
\arguments{
\item{file}{Path to a file containing data from RMSC best track: https://www.jma.go.jp/jma/jma-eng/jma-center/rsmc-hp-pub-eg/besttrack.html}
}
\value{
A data.frame containing columns timestep, longitude, latitude, rel_vor_850
}
\description{
read_rsmc_besttrack
}
