% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_track.r
\name{read_track}
\alias{read_track}
\title{read_track}
\usage{
read_track(file)
}
\arguments{
\item{file}{Path to a file containing data from Hodges' TRACK algorithm}
}
\value{
A data.frame containing columns timestep, longitude, latitude, rel_vor_850
}
\description{
read_track
}
