% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_track.r
\name{track_to_sf_lines}
\alias{track_to_sf_lines}
\title{track_to_sf_lines
Convert the data.frame output of read_track to a SimpleFeatures LINES object
for use with GIS functionality}
\usage{
track_to_sf_lines(track_df)
}
\arguments{
\item{track_df}{data.frame, output of read_track()}
}
\value{
sf (LINES)
}
\description{
track_to_sf_lines
Convert the data.frame output of read_track to a SimpleFeatures LINES object
for use with GIS functionality
}
