% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_track.r
\name{track_to_sf_points}
\alias{track_to_sf_points}
\title{track_to_sf_points
Convert the data.frame output of read_track to a SimpleFeatures POINT object
for use with GIS functionality}
\usage{
track_to_sf_points(track_df)
}
\arguments{
\item{track_df}{data.frame, output of read_track()}
}
\value{
sf (POINTS)
}
\description{
track_to_sf_points
Convert the data.frame output of read_track to a SimpleFeatures POINT object
for use with GIS functionality
}
